Open Note: parse built-in metadata and replace if embedded does not match with note's metadata (fix #112)
This solves the problem that changing metadata in other clients
does not remove the comment with the built-in markdown
making the embedded metadata header inconsistent with the real metadata.

Here we check if the header is consistent otherwise we replace it.
We do not always replace it to support workflows that
keep custom metadata in the header
but never change the note's metadata in other clients.